uselessfs: introduce the uselessfs FUSE overlay
The purpose of uselessfs is to serve as a virtual file system representation
of IPC resources and arguments (whereas FIFO control commands essentially
serve as the "method calls"). The idea is that one can write values to
a file named after an action or property (e.g. kill, kill-signal, start),
then send the pipe a control command of "kill", which will trigger a read
of the kill file and pass its contents as a parameter to the relevant
function.

The files are purely synthetic, the VFS is inherently introspectable,
can be configured via mount flags and fuse.conf, and can be transported.
FUSE is also portable.

Right now, this is just a "hello world" stub adapted from bbfs.